refactor : main package 정리 작업, Main 내 setTablayout 생성, splashActivity 리팩토링
