+++ app/src/main/java/com/example/earlybirdy/main/MainViewPagerAdapter.kt
@@ -1,8 +1,9 @@
-package com.example.earlybirdy
+package com.example.earlybirdy.main
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.example.earlybirdy.R
 import com.example.earlybirdy.board.BoardFragment
 import com.example.earlybirdy.home.HomeFragment
 import com.example.earlybirdy.my_page.MyPageFragment
